Crate iota_sdk_types

Core type definitions for the IOTA blockchain.

IOTA is a next-generation smart contract platform with high throughput, low latency, and an asset-oriented programming model powered by the Move programming language. This crate provides type definitions for working with the data that makes up the IOTA blockchain.

Feature flags

This library uses a set of feature flags to reduce the number of dependencies and amount of compiled code. By default, no features are enabled which allows one to enable a subset specifically for their use case. Below is a list of the available feature flags.

serde: Enables support for serializing and deserializing types to/from BCS utilizing serde library. Note: JSON serialization is NOT guaranteed to match the IOTA monorepo’s JSON-RPC format.

rand: Enables support for generating random instances of a number of types via the rand library.

hash: Enables support for hashing, which is required for deriving addresses and calculating digests for various types.

proptest: Enables support for the proptest library by providing implementations of proptest::arbitrary::Arbitrary for many types.

BCS

BCS is the serialization format used to represent the state of the blockchain and is used extensively throughout the IOTA ecosystem. In particular the BCS format is leveraged because it “guarantees canonical serialization, meaning that for any given data type, there is a one-to-one correspondence between in-memory values and valid byte representations.” One benefit of this property of having a canonical serialized representation is to allow different entities in the ecosystem to all agree on how a particular type should be interpreted and more importantly define a deterministic representation for hashing and signing.

This library strives to guarantee that the types defined are fully BCS-compatible with the data that the network produces. The one caveat to this would be that as the IOTA protocol evolves, new type variants are added and older versions of this library may not support those newly added variants. The expectation is that the most recent release of this library will support new variants and types as they are released to IOTA’s testnet network.

The BCS serialized form of every type in this crate is specified in ABNF notation, as described by RFC-5234, in bcs-schema.abnf. In addition to the format itself, some types have an extra layer of verification and may impose additional restrictions on valid byte representations above and beyond those already provided by BCS. In these instances the documentation for those types will clearly specify these additional restrictions.

Re-exports

pub use crypto::SignatureFromBytesError;serde

pub use crypto::Bls12381PublicKey;

pub use crypto::Bls12381Signature;

pub use crypto::Ed25519PublicKey;

pub use crypto::Ed25519Signature;

pub use crypto::HashingIntentScope;

pub use crypto::INTENT_PREFIX_LENGTH;

pub use crypto::Intent;

pub use crypto::IntentAppId;

pub use crypto::IntentError;

pub use crypto::IntentMessage;

pub use crypto::IntentScope;

pub use crypto::IntentVersion;

pub use crypto::MoveAuthenticator;

pub use crypto::MoveAuthenticatorV1;

pub use crypto::MultisigAggregatedSignature;

pub use crypto::MultisigCommittee;

pub use crypto::MultisigError;

pub use crypto::MultisigMember;

pub use crypto::MultisigMemberSignature;

pub use crypto::PasskeyAuthenticator;

pub use crypto::PasskeyPublicKey;

pub use crypto::PersonalMessage;

pub use crypto::PublicKey;

pub use crypto::PublicKeyError;

pub use crypto::PublicKeyExt;

pub use crypto::Secp256k1PublicKey;

pub use crypto::Secp256k1Signature;

pub use crypto::Secp256r1PublicKey;

pub use crypto::Secp256r1Signature;

pub use crypto::SignatureScheme;

pub use crypto::SignatureSchemeError;

pub use crypto::SimpleSignature;

pub use crypto::UserSignature;

pub use effects::BalanceChange;

pub use effects::ChangedObject;

pub use effects::DeriveChangesError;

pub use effects::IdOperation;

pub use effects::InputSharedObject;

pub use effects::ObjectChange;

pub use effects::ObjectIn;

pub use effects::ObjectOut;

pub use effects::ObjectRemoveKind;

pub use effects::TransactionEffects;

pub use effects::TransactionEffectsV1;

pub use effects::UnchangedSharedKind;

pub use effects::UnchangedSharedObject;

pub use effects::WriteKind;

pub use framework::Coin;

pub use framework::CoinFromObjectError;

pub use iota_names::error::IotaNamesError;

pub use version::Version;

pub use version::VersionError;

Modules

crypto

effects

framework

Rust definitions of move/iota framework types.

hashhash

iota_names

version

Structs

Address

Unique identifier for an Account on the IOTA blockchain.

CanceledTransaction

A transaction that was canceled

CertificateDigest

The digest of a certificate.

ChangeEpoch

System transaction used to change the epoch

ChangeEpochV2

System transaction used to change the epoch

ChangeEpochV3

BCS

ChangeEpochV4

BCS

CheckpointContentsDigest

The digest of a CheckpointContents.

CheckpointContentsV1

CheckpointContents are the transactions included in an upcoming checkpoint. They must have already been causally ordered. Since the causal order algorithm is the same among validators, we expect all honest validators to come up with the same order for each checkpoint content.

CheckpointData

BCS

CheckpointDigest

The digest of a CheckpointSummary.

CheckpointSummary

A header for a Checkpoint on the IOTA blockchain.

CheckpointTransaction

BCS

CheckpointTransactionInfo

Transaction information committed to in a checkpoint

ConsensusCommitDigest

The digest of a consensus commit.

ConsensusCommitPrologueV1

V1 of the consensus commit prologue system transaction

DenyRuleSet

A complete set of transaction deny rules.

Digest

A 32-byte Blake2b256 hash output.

EffectsAuxDataDigest

The digest of the auxiliary data associated with TransactionEffects.

EndOfEpochData

Data, which when included in a CheckpointSummary, signals the end of an Epoch.

Event

An event

GasCostSummary

Summary of gas charges.

GasPayment

Payment information for executing a transaction

GenesisObject

An object part of the initial chain state

GenesisTransaction

The genesis transaction

Identifier

A move identifier

MakeMoveVector

Command to build a move vector out of a set of individual elements

MergeCoins

Command to merge multiple coins of the same type into a single coin

MisbehaviorReportDigest

The digest of a misbehavior report.

MoveAuthenticatorDigest

The digest of a MoveAuthenticator.

MoveCall

Command to call a move function

MoveLocation

Location in move bytecode where an error occurred

MoveObjectType

A StructTag with optimized BCS serialization for object types.

MovePackage

A move package

MovePackageData

Type corresponding to the output of iota move build --dump-bytecode-as-base64

MoveStruct

A move struct

MoveStructContentsError

Error returned when MoveStruct contents are too short to contain an ObjectId.

Object

An object on the IOTA blockchain

ObjectDigest

The digest of an Object.

ObjectId

An ObjectId is a 32-byte identifier used to uniquely identify an object on the IOTA blockchain.

ObjectReference

Reference to an object

ObjectVersion

An ObjectId paired with one of that object’s versions.

OwnedObjectReference

An ObjectReference paired with the owner the object has at that version.

ProgrammableTransaction

A user transaction

Publish

Command to publish a new move package

RandomnessRound

Round number of generated randomness.

RandomnessStateUpdate

Randomness update

SenderSignedDataDigest

The digest of SenderSignedData.

SenderSignedTransaction

A SignedTransaction in its intent-message serialized form.

SharedObjectReference

A shared object input to a programmable transaction

SignedCheckpointSummary

BCS

SignedTransaction

BCS

SignerBitmapErrorserde

Error returned when a signer bitmap cannot be read.

SplitCoins

Command to split a single coin object into multiple coins

StructTag

Type information for a move struct

SystemPackage

BCS

TransactionDenyRulesUpdate

Update of the on-chain transaction deny rules.

TransactionDigest

The digest of a Transaction.

TransactionEffectsDigest

The digest of TransactionEffects.

TransactionEvents

Events emitted during the successful execution of a transaction

TransactionEventsDigest

The digest of TransactionEvents.

TransactionV1

BCS

TransferObjects

Command to transfer ownership of a set of objects to an address

TypeOrigin

Stores the origin of a data type where it first appeared in the version chain. A data type is identified by the name of the module and the name of the struct/enum in combination.

Upgrade

Command to upgrade an already published package

UpgradeInfo

Upgraded package info for MovePackage’s linkage_table.

ValidatorAggregatedSignature

An aggregated signature from multiple Validators.

ValidatorCommittee

The Validator Set for a particular epoch.

ValidatorCommitteeMember

A member of a Validator Committee

ValidatorSignature

A signature from a Validator

VersionAssignment

Object version assignment from consensus

Enums

AddressParseError

Argument

An argument to a programmable transaction command

CheckpointCommitment

A commitment made by a checkpoint.

CheckpointContents

The committed to contents of a checkpoint.

Command

A single command in a programmable transaction.

CommandArgumentError

An error with an argument to a command

ConsensusDeterminedVersionAssignments

BCS

DigestParseError

EndOfEpochTransactionKind

Operation run at the end of an epoch

ExecutionError

An error that can occur during the execution of a transaction

ExecutionStatus

The status of an executed Transaction

FromBase64Errorserde

Error returned when decoding a type from a base64-encoded string of its BCS bytes.

Input

An input to a user transaction

ObjectData

Object data, either a package or struct

ObjectType

Type of an IOTA object

Owner

Enum of different types of ownership for an object.

PackageUpgradeError

An error with a upgrading a package

Transaction

Transaction

TransactionExpiration

A TTL for a transaction

TransactionKind

Transaction type

TypeArgumentError

An error with a type argument

TypeParseError

TypeTag

Type of a move value

UpgradePolicy

Rust representation of upgrade policy constants in iota::package.

ValidatorCommitteeError

An error returned when a ValidatorCommittee is not well-formed.

Constants

MAX_IDENTIFIER_LENGTH

Maximum length in bytes of a Move Identifier.

MAX_TYPE_TAG_NESTING

Maximum nesting depth allowed when parsing a TypeTag.

Type Aliases

CheckpointSequenceNumber

CheckpointTimestamp

EpochId

ProtocolVersion

SigningDigest

StakeUnit