iota_sdk_types

Module checkpoint

Structs

CheckpointContentsV1

CheckpointContents are the transactions included in an upcoming checkpoint. They must have already been causally ordered. Since the causal order algorithm is the same among validators, we expect all honest validators to come up with the same order for each checkpoint content.

CheckpointData

CheckpointSummary

A header for a Checkpoint on the IOTA blockchain.

CheckpointTransaction

CheckpointTransactionInfo

Transaction information committed to in a checkpoint

EndOfEpochData

Data, which when included in a CheckpointSummary, signals the end of an Epoch.

SignedCheckpointSummary

Enums

CheckpointCommitment

A commitment made by a checkpoint.

CheckpointContents

The committed to contents of a checkpoint.

Type Aliases

CheckpointSequenceNumber

CheckpointTimestamp

EpochId

ProtocolVersion

StakeUnit