iota_sdk_types

Module effects

Re-exports

pub use changes::BalanceChange;

pub use changes::DeriveChangesError;

pub use changes::ObjectChange;

Modules

changes

Derivation of a transaction’s balance and object changes from its effects and the objects it read and wrote.

Structs

ChangedObject

Input/output state of an object that was changed during execution

TransactionEffectsV1

Version 1 of TransactionEffects

UnchangedSharedObject

A shared object that wasn’t changed during execution

Enums

IdOperation

Defines what happened to an ObjectId during execution

InputSharedObject

A shared object an executed transaction took as input.

ObjectIn

State of an object prior to execution

ObjectOut

State of an object after execution

ObjectRemoveKind

Why an object is no longer in the store after a transaction.

TransactionEffects

The output or effects of executing a transaction

UnchangedSharedKind

Type of unchanged shared object

WriteKind

How an object came to be in the store after a transaction wrote it.