iota_sdk_types

Enum MultisigMemberSignature

pub enum MultisigMemberSignature {
    Ed25519(Ed25519Signature),
    Secp256k1(Secp256k1Signature),
    Secp256r1(Secp256r1Signature),
    ZkLogin(Box<ZkLoginAuthenticator>),
}

A signature from a member of a multisig committee.

BCS

The BCS serialized form for this type is defined by the following ABNF:

multisig-member-signature = ed25519-multisig-member-signature /
                            secp256k1-multisig-member-signature /
                            secp256r1-multisig-member-signature /
                            zklogin-multisig-member-signature

ed25519-multisig-member-signature   = %x00 ed25519-signature
secp256k1-multisig-member-signature = %x01 secp256k1-signature
secp256r1-multisig-member-signature = %x02 secp256r1-signature
zklogin-multisig-member-signature   = %x03 zklogin-authenticator

Variants

Ed25519(Ed25519Signature)

Secp256k1(Secp256k1Signature)

Secp256r1(Secp256r1Signature)

ZkLogin(Box<ZkLoginAuthenticator>)

Implementations

impl MultisigMemberSignature

pub fn is_ed25519(&self) -> bool

pub fn as_ed25519(&self) -> &Ed25519Signature

pub fn as_ed25519_opt(&self) -> Option<&Ed25519Signature>

pub fn into_ed25519_opt(self) -> Option<Ed25519Signature>

pub fn into_ed25519(self) -> Ed25519Signature

pub fn is_secp256k1(&self) -> bool

pub fn as_secp256k1(&self) -> &Secp256k1Signature

pub fn as_secp256k1_opt(&self) -> Option<&Secp256k1Signature>

pub fn into_secp256k1_opt(self) -> Option<Secp256k1Signature>

pub fn into_secp256k1(self) -> Secp256k1Signature

pub fn is_secp256r1(&self) -> bool

pub fn as_secp256r1(&self) -> &Secp256r1Signature

pub fn as_secp256r1_opt(&self) -> Option<&Secp256r1Signature>

pub fn into_secp256r1_opt(self) -> Option<Secp256r1Signature>

pub fn into_secp256r1(self) -> Secp256r1Signature

pub fn is_zklogin(&self) -> bool

pub fn as_zklogin(&self) -> &ZkLoginAuthenticator

pub fn as_zklogin_opt(&self) -> Option<&ZkLoginAuthenticator>

pub fn into_zklogin_opt(self) -> Option<ZkLoginAuthenticator>

pub fn into_zklogin(self) -> ZkLoginAuthenticator