iota_sdk_types

Module move_package

Structs

MovePackage

A move package

MovePackageData

Type corresponding to the output of iota move build --dump-bytecode-as-base64

TypeOrigin

Stores the origin of a data type where it first appeared in the version chain. A data type is identified by the name of the module and the name of the struct/enum in combination.

UpgradeInfo

Upgraded package info for MovePackage’s linkage_table.

Enums

UpgradePolicy

Rust representation of upgrade policy constants in iota::package.