iota_sdk_types

Module object

Structs

GenesisObject

An object part of the initial chain state

MoveObjectType

A StructTag with optimized BCS serialization for object types.

MoveStruct

A move struct

MoveStructContentsError

Error returned when MoveStruct contents are too short to contain an ObjectId.

Object

An object on the IOTA blockchain

ObjectReference

Reference to an object

ObjectVersion

An ObjectId paired with one of that object’s versions.

OwnedObjectReference

An ObjectReference paired with the owner the object has at that version.

Enums

ObjectData

Object data, either a package or struct

ObjectType

Type of an IOTA object

Owner

Enum of different types of ownership for an object.