iota_sdk_types

Module transaction

Structs

CanceledTransaction

A transaction that was canceled

ChangeEpoch

System transaction used to change the epoch

ChangeEpochV2

System transaction used to change the epoch

ChangeEpochV3

ChangeEpochV4

ConsensusCommitPrologueV1

V1 of the consensus commit prologue system transaction

DenyRuleSet

A complete set of transaction deny rules.

GasPayment

Payment information for executing a transaction

GenesisTransaction

The genesis transaction

MakeMoveVector

Command to build a move vector out of a set of individual elements

MergeCoins

Command to merge multiple coins of the same type into a single coin

MoveCall

Command to call a move function

ProgrammableTransaction

A user transaction

Publish

Command to publish a new move package

RandomnessRound

Round number of generated randomness.

RandomnessStateUpdate

Randomness update

SenderSignedTransaction

A SignedTransaction in its intent-message serialized form.

SharedObjectReference

A shared object input to a programmable transaction

SignedTransaction

SplitCoins

Command to split a single coin object into multiple coins

SystemPackage

TransactionDenyRulesUpdate

Update of the on-chain transaction deny rules.

TransactionV1

TransferObjects

Command to transfer ownership of a set of objects to an address

Upgrade

Command to upgrade an already published package

VersionAssignment

Object version assignment from consensus

Enums

Argument

An argument to a programmable transaction command

Command

A single command in a programmable transaction.

ConsensusDeterminedVersionAssignments

EndOfEpochTransactionKind

Operation run at the end of an epoch

Input

An input to a user transaction

Transaction

Transaction

TransactionExpiration

A TTL for a transaction

TransactionKind

Transaction type