iota_sdk_types::object

Struct Object

pub struct Object {
    pub data: ObjectData,
    pub owner: Owner,
    pub previous_transaction: TransactionDigest,
    pub storage_rebate: u64,
}

An object on the IOTA blockchain

BCS

The BCS serialized form for this type is defined by the following ABNF:

object = object-data owner digest u64

Fields
data: ObjectData

The meat of the object

owner: Owner

The owner that unlocks this object

previous_transaction: TransactionDigest

The digest of the transaction that created or last mutated this object

storage_rebate: u64

The amount of IOTA we would rebate if this object gets deleted. This number is re-calculated each time the object is mutated based on the present storage gas price.

Implementations

impl Object

pub fn digest(&self) -> ObjectDigest

Available on crate features hash and serde only.

Calculate the digest of this Object

This is done by hashing the BCS bytes of this Object prefixed with a salt.

impl Object

pub fn new( data: ObjectData, owner: Owner, previous_transaction: TransactionDigest, storage_rebate: u64, ) -> Self

Build an object

pub fn id(&self) -> ObjectId

Return this object’s id

pub fn object_ref(&self) -> ObjectReference

Available on crate features hash and serde only.

Return this object’s reference

pub fn version(&self) -> Version

Return this object’s version

pub fn object_type(&self) -> ObjectType

Return this object’s type

pub fn as_opt_struct(&self) -> Option<&MoveStruct>

Try to interpret this object as a move struct

pub fn as_struct(&self) -> &MoveStruct

Interpret this object as a move struct

pub fn as_opt_package(&self) -> Option<&MovePackage>

Try to interpret this object as a move package

pub fn as_package(&self) -> &MovePackage

Interpret this object as a move package

pub fn owner(&self) -> &Owner

Return this object’s owner

pub fn data(&self) -> &ObjectData

Return this object’s data

pub fn previous_transaction(&self) -> TransactionDigest

Return the digest of the transaction that last modified this object

pub fn storage_rebate(&self) -> u64

Return the storage rebate locked in this object

Storage rebates are credited to the gas coin used in a transaction that deletes this object.

pub fn to_rust<'de, T: Deserialize<'de>>( &'de self, ) -> Result<T, Box<dyn Error + Send + Sync>>

Available on crate feature serde only.

pub fn is_immutable(&self) -> bool

Returns true if the object is immutable.

pub fn is_address_owned(&self) -> bool

Returns true if the object is owned by an address.

pub fn is_child_object(&self) -> bool

Returns true if the object is owned by another object.

pub fn is_shared(&self) -> bool

Returns true if the object is shared.

pub fn is_package(&self) -> bool

Returns true if this object is a Move package rather than a Move value.

pub fn is_system_package(&self) -> bool

Returns true if the object is a system package.

pub fn struct_tag(&self) -> Option<StructTag>

Returns the struct tag of this object if it is a Move struct.

pub fn is_gas_coin(&self) -> bool

Returns true if this object is a gas coin.

pub fn opt_coin_type(&self) -> Option<&TypeTag>

Returns the coin’s type parameter if this object is a coin.

pub fn single_owner(&self) -> Option<Address>

Returns the address of the single owner of this object (address- or object-owned), or None if it is shared or immutable.

pub fn set_owner(&mut self, new_owner: Address)

Sets the owner of this object to new_owner.

impl Object

pub fn to_bcs(&self) -> Vec<u8>

Available on crate feature serde only.

Serialize this Object as a Vec<u8> of BCS bytes.

pub fn to_base64(&self) -> String

Available on crate feature serde only.

Serialize this Object as a base64-encoded string of its BCS bytes.

pub fn from_bcs(bytes: &[u8]) -> Result<Self, Error>

Available on crate feature serde only.

Deserialize a Object from BCS bytes.

pub fn from_base64(bytes: &str) -> Result<Self, Error>

Available on crate feature serde only.

Deserialize a Object from a base64-encoded string of its BCS bytes.